Move card-scanning classes to old package.
COSDK-1195
