+++ card/src/main/java/com/adyen/checkout/card/old/internal/util/CardScannerWrapper.kt
@@ -11,8 +11,8 @@ package com.adyen.checkout.card.old.internal.util
 import android.content.Context
 import android.content.Intent
 import androidx.fragment.app.Fragment
-import com.adyen.checkout.card.scanning.AdyenCardScanner
-import com.adyen.checkout.card.scanning.AdyenCardScannerResult
+import com.adyen.checkout.card.scanning.old.AdyenCardScanner
+import com.adyen.checkout.card.scanning.old.AdyenCardScannerResult
 import com.adyen.checkout.core.old.Environment
 import com.adyen.checkout.core.old.internal.util.runCompileOnly
